+++ spec/spec_helper.rb
@@ -1,4 +1,7 @@
 # encoding: utf-8
+require 'codeclimate-test-reporter'
+CodeClimate::TestReporter.start
+
 require 'rubygems'
 require 'spork'