+++ lib/ndd/rspec/matchers.rb
@@ -1,6 +1,6 @@
 # encoding: utf-8
-require_relative 'matchers/be_reverse_sorted_by'
-require_relative 'matchers/be_sorted_by'
+require 'ndd/rspec/matchers/be_reverse_sorted_by'
+require 'ndd/rspec/matchers/be_sorted_by'
 
 
 module RSpec::Matchers