+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * updated Ruby version to 2.1
 * updated RSpec version to 3.1
 * updated version of all development dependencies
+* added CodeClimate test coverage
 
 ## Version 0.1.0